backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
python-multipart==0.0.20
pillow==11.0.0
numpy==2.1.3
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
psycopg2-binary>=2.9.9
boto3>=1.34.0
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.56.0
openai>=1.40.0
# YOLO (valgfritt): pip install ultralytics
ultralytics>=8.3.0
